Renaming new alamopy wrapper to alamopy
